Arreglos de conflictos luego del Pull de develop 3
