+++ app/src/main/java/com/example/finalproject/ui/home/adapter/ProductsAdapter.kt
@@ -5,9 +5,8 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.example.finalproject.data.dto.response.Product
 import com.example.finalproject.databinding.ItemRvHomeProductsBinding
-import com.squareup.picasso.Picasso
 
-class ProductsAdapter : RecyclerView.Adapter<ProductsAdapter.ProductViewHolder>() {
+class ProductsAdapter : RecyclerView.Adapter<ProductViewHolder>() {
 
     private val productList = mutableListOf<Product>()
 
@@ -28,15 +27,6 @@ class ProductsAdapter : RecyclerView.Adapter<ProductsAdapter.ProductViewHolder>(
     }
 
     override fun getItemCount(): Int = productList.size
-
-    class ProductViewHolder(private val binding: ItemRvHomeProductsBinding) :
-        RecyclerView.ViewHolder(binding.root) {
-        fun bind(product: Product) {
-            binding.nameRecyclerProduct.text = product.name
-            binding.tvRecyclerPrice.text = "${product.currency}${product.price}"
-            Picasso.get().load(product.image).into(binding.ivRecyclerProduct)
-        }
-    }
 }
